tbl_login_info
arrranged login so that admin lecturer and student can access certain parts of the website.

Came with the idea that students can input their link.

Lecturers can choose their favourite students (for the portfolio so they best appear in front).  Lecturer may also input news and when he's absent.

Admin inputs the rest of the data.